requirements.txt
asttokens==2.4.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.0
comtypes==1.2.0
debugpy==1.8.0
decorator==5.1.1
element-manager==1.0.4
exceptiongroup==1.2.0
executing==2.0.1
Faker==20.1.0
h11==0.14.0
idna==3.6
importlib-metadata==7.0.0
ipykernel==6.27.1
ipython==8.12.3
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
mouse==0.7.1
MouseInfo==0.1.3
mss==9.0.1
nest-asyncio==1.5.8
numpy==1.19.3
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
PySide2==5.15.2.1
PySimpleGUI==4.60.5
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pywin32==306
pywinauto==0.6.8
pyzmq==25.1.2
requests==2.31.0
robotframework==6.1.1
rpapy==1.0.2
selenium==4.16.0
shiboken2==5.15.2.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.1.0
wcwidth==0.2.12
webdriver-manager==4.0.1
win10toast==0.9
wsproto==1.2.0
zipp==3.17.0
